[feat] 포트폴리오 종목 조회 구현 (#49)
* #10 feat: 포트폴리오 종목 목록 조회 구현

* #10 feat: @JsonUnWraaped 적용

+++ fineAnts_app/src/test/java/codesquad/fineants/domain/portfolio_stock/PortfolioHoldingTest.java
@@ -67,10 +67,10 @@ class PortfolioHoldingTest {
 		portFolioHolding.addTradeHistory(tradeHistory2);
 
 		// when
-		long result = portFolioHolding.calculateAverageCostPerShare();
+		Double result = portFolioHolding.calculateAverageCostPerShare();
 
 		// then
-		assertThat(result).isEqualTo(10000L);
+		assertThat(result).isEqualTo(10000.0);
 	}
 
 	@DisplayName("한 종목의 총 손익을 계산한다")
